// src/shared/api/models/getApiDoctorsPortalDoctors200DoctorsItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiDoctorsPortalDoctors200DoctorsItemPhoneNumber } from './getApiDoctorsPortalDoctors200DoctorsItemPhoneNumber';
import type { GetApiDoctorsPortalDoctors200DoctorsItemPhoneCountryCode } from './getApiDoctorsPortalDoctors200DoctorsItemPhoneCountryCode';
import type { GetApiDoctorsPortalDoctors200DoctorsItemPhoneCountryIsoCode } from './getApiDoctorsPortalDoctors200DoctorsItemPhoneCountryIsoCode';
import type { GetApiDoctorsPortalDoctors200DoctorsItemUpdatedAt } from './getApiDoctorsPortalDoctors200DoctorsItemUpdatedAt';
import type { GetApiDoctorsPortalDoctors200DoctorsItemLastConnectionStatus } from './getApiDoctorsPortalDoctors200DoctorsItemLastConnectionStatus';

export type GetApiDoctorsPortalDoctors200DoctorsItem = {
    uuid: string;
    email: string;
    firstName: string;
    lastName: string;
    phoneNumber: GetApiDoctorsPortalDoctors200DoctorsItemPhoneNumber;
    phoneCountryCode: GetApiDoctorsPortalDoctors200DoctorsItemPhoneCountryCode;
    phoneCountryIsoCode: GetApiDoctorsPortalDoctors200DoctorsItemPhoneCountryIsoCode;
    isActive: boolean;
    createdAt: string;
    updatedAt: GetApiDoctorsPortalDoctors200DoctorsItemUpdatedAt;
    lastConnectionStatus: GetApiDoctorsPortalDoctors200DoctorsItemLastConnectionStatus;
    isLinked?: boolean;
    isInvited?: boolean;
    linkedDate?: string;
};